# backend/ent/schema/staff.go
package schema

import (
	"github.com/facebookincubator/ent"
	"github.com/facebookincubator/ent/schema/edge"
	"github.com/facebookincubator/ent/schema/field"
)

// Staff holds the schema definition for the Staff entity.
type Staff struct {
	ent.Schema
}

// Fields of the Staff.
func (Staff) Fields() []ent.Field {
	return []ent.Field{
		field.String("staff_email").Unique(),
		field.String("staff_name").NotEmpty(),
		field.String("staff_password").NotEmpty(),
	}
}

// Edges of the Staff.
func (Staff) Edges() []ent.Edge {
	return []ent.Edge{
		edge.To("staff_predicament", Predicament.Type).StorageKey(edge.Column("staffid")),
	}
}